qmake: improve and optimize pkg-config usage
Make use of the qmake feature for finding packages using pkg-config,
and using their results for include paths and libraries.
Also, optimize the way the version of poppler is read: query pkg-config
only once at qmake time, splitting the bits in variables added as
CFLAGS; this way, we avoids spawning 99 times (!) pkg-config and cut
during build.

All of this is done on Unix systems only, as pkg-config is available
only there.