First support for building with Qt5
When the Qt version is greater than 4 (so 5 at least), require the
QtWidgets module as well, which is where QWidget-based stuff lives.
Temporarily enable deprecated methods in Qt5, so there are no code
changes required.

Also pick the right poppler version, poppler-qt4 for Qt4 and
poppler-qt5 for Qt5, including the right header depending on the
version.

These changes allow katarakt to build and link, and it seems to be
working fine at a quick glance.